Integration testing: common structure for all adapters.
